fixed structfield comparison when dataType is array and ignore_nullability=True
