+++ chispa/schema_comparer.py
@@ -1,7 +1,7 @@
 from chispa.prettytable import PrettyTable
 from chispa.bcolors import *
 import chispa.six as six
-from chispa.structfield_comparer import are_structfields_equal
+from chispa.structfield_comparer import are_structfields_equal, check_type_equal_ignore_nullable
 
 
 class SchemasNotEqualError(Exception):
@@ -52,17 +52,4 @@ def are_schemas_equal_ignore_nullable(s1, s2):
     return True
 
 
-def check_type_equal_ignore_nullable(sf1, sf2):
-    """Checks StructField data types ignoring nullables.
 
-    Handles array element types also.
-    """
-    dt1, dt2 = sf1.dataType, sf2.dataType
-    if dt1.typeName() == dt2.typeName():
-        # Account for array types by inspecting elementType.
-        if dt1.typeName() == 'array':
-            return dt1.elementType == dt2.elementType
-        else:
-            return True
-    else:
-        return False
